# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Neuralix Dashboard API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /guilds/{guildId}:
    get:
      operationId: getGuild
      tags: [guilds]
      summary: Get guild details
      parameters:
        - name: guildId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Guild details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GuildDetail"
        "404":
          description: Not found
  /guilds/{guildId}/stats:
    get:
      operationId: getGuildStats
      tags: [guilds]
      summary: Get guild statistics dashboard summary
      parameters:
        - name: guildId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Guild stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GuildStats"
  /guilds/{guildId}/bot-status:
    get:
      operationId: getGuildBotStatus
      tags: [guilds]
      summary: Check if bot is in guild
      parameters:
        - name: guildId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Bot status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BotStatus"

  # ANNOUNCEMENTS
  /announcements:
    get:
      operationId: getAnnouncements
      tags: [announcements]
      summary: Get public announcements
      responses:
        "200":
          description: List of announcements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Announcement"
    post:
      operationId: createAnnouncement
      tags: [announcements]
      summary: Create announcement (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnnouncementInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Announcement"
  /announcements/{id}:
    patch:
      operationId: updateAnnouncement
      tags: [announcements]
      summary: Update announcement
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnnouncementUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Announcement"
    delete:
      operationId: deleteAnnouncement
      tags: [announcements]
      summary: Delete announcement
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # WELCOME
...
  /guilds/{guildId}/antiraid:
    get:
      operationId: getAntiraidConfig
      tags: [antiraid]
      summary: Get AntiRaid configuration
      parameters:
        - name: guildId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: AntiRaid config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AntiraidConfig"
    put:
      operationId: updateAntiraidConfig
      tags: [antiraid]
      summary: Update AntiRaid configuration
      parameters:
        - name: guildId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AntiraidConfigInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AntiraidConfig"
  /guilds/{guildId}/antiraid/stats:
    get:
      operationId: getAntiraidStats
      tags: [antiraid]
      summary: Get AntiRaid statistics
      parameters:
        - name: guildId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: AntiRaid stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AntiraidStats"

  # LOGS
  /guilds/{guildId}/logs/config:
    get:
      operationId: getLogsConfig
      tags: [logs]
      summary: Get logs configuration
      parameters:
        - name: guildId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Logs config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LogsConfig"
    put:
      operationId: updateLogsConfig
      tags: [logs]
      summary: Update logs configuration
      parameters:
        - name: guildId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LogsConfigInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LogsConfig"
  /guilds/{guildId}/logs:
    get:
      operationId: getGuildLogs
      tags: [logs]
      summary: Get logs for a guild
      parameters:
        - name: guildId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Log entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LogEntry"

  # BLACKLIST
  /blacklist:
    get:
      operationId: getBlacklist
      tags: [blacklist]
      summary: Get global blacklist (owner only)
      responses:
        "200":
          description: Blacklist entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BlacklistEntry"
    post:
      operationId: addToBlacklist
      tags: [blacklist]
      summary: Add user to blacklist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BlacklistInput"
      responses:
        "201":
          description: Added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlacklistEntry"
  /blacklist/{userId}:
    delete:
      operationId: removeFromBlacklist
      tags: [blacklist]
      summary: Remove user from blacklist
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Removed

  # BACKUPS
  /guilds/{guildId}/backups:
    get:
      operationId: getBackups
      tags: [backups]
      summary: Get backups for guild
      parameters:
        - name: guildId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Backups list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Backup"
    post:
      operationId: createBackup
      tags: [backups]
      summary: Create a backup
      parameters:
        - name: guildId
          in: path
          required: true
          schema:
            type: string
      responses:
        "201":
          description: Backup created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Backup"
  /guilds/{guildId}/backups/{backupId}/restore:
    post:
      operationId: restoreBackup
      tags: [backups]
      summary: Restore a backup
      parameters:
        - name: guildId
          in: path
          required: true
          schema:
            type: string
        - name: backupId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Restored

  # PREMIUM
...
components:
  schemas:
...
    GuildDetail:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        icon:
          type: ["string", "null"]
        memberCount:
          type: integer
        onlineMemberCount:
          type: integer
        botPresent:
          type: boolean
        premiumTier:
          type: integer
        openTickets:
          type: integer
        verifiedMembers:
          type: integer
        antiraidEnabled:
          type: boolean
        premiumActive:
          type: boolean
      required: [id, name, memberCount, onlineMemberCount, botPresent, premiumTier, openTickets, verifiedMembers, antiraidEnabled, premiumActive]
    GuildStats:
      type: object
      properties:
        guildId:
          type: string
        memberCount:
          type: integer
        onlineCount:
          type: integer
        openTickets:
          type: integer
        closedTickets:
          type: integer
        verifiedMembers:
          type: integer
        antiraidDetections:
          type: integer
        recentLogs:
          type: integer
        backupsCount:
          type: integer
      required: [guildId, memberCount, onlineCount, openTickets, closedTickets, verifiedMembers, antiraidDetections, recentLogs, backupsCount]
    BotStatus:
      type: object
      properties:
        present:
          type: boolean
        addBotUrl:
          type: string
      required: [present, addBotUrl]
    Announcement:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        type:
          type: string
        published:
          type: boolean
        scheduledAt:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, title, content, type, published, createdAt, updatedAt]
    AnnouncementInput:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        type:
          type: string
        published:
          type: boolean
        scheduledAt:
          type: ["string", "null"]
      required: [title, content, type]
    AnnouncementUpdate:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        type:
          type: string
        published:
          type: boolean
        scheduledAt:
          type: ["string", "null"]
...
    AntiraidConfig:
      type: object
      properties:
        id:
          type: integer
        guildId:
          type: string
        enabled:
          type: boolean
        antiAlt:
          type: boolean
        antiAltMinAge:
          type: integer
        antiBot:
          type: boolean
        antiBotWhitelist:
          type: array
          items:
            type: string
        antiSpam:
          type: boolean
        antiSpamLimit:
          type: integer
        antiLinks:
          type: boolean
        allowedDomains:
          type: array
          items:
            type: string
        blockedDomains:
          type: array
          items:
            type: string
        antiMassMention:
          type: boolean
        massMentionLimit:
          type: integer
        antiWebhook:
          type: boolean
        antiChannelCreate:
          type: boolean
        antiChannelDelete:
          type: boolean
        antiChannelUpdate:
          type: boolean
        antiRoleCreate:
          type: boolean
        antiRoleDelete:
          type: boolean
        antiRoleUpdate:
          type: boolean
        antiEmojiCreate:
          type: boolean
        antiEmojiDelete:
          type: boolean
        antiBanMass:
          type: boolean
        antiKickMass:
          type: boolean
        antiNuke:
          type: boolean
      required: [id, guildId, enabled, antiAlt, antiAltMinAge, antiBot, antiBotWhitelist, antiSpam, antiSpamLimit, antiLinks, allowedDomains, blockedDomains, antiMassMention, massMentionLimit, antiWebhook, antiChannelCreate, antiChannelDelete, antiChannelUpdate, antiRoleCreate, antiRoleDelete, antiRoleUpdate, antiEmojiCreate, antiEmojiDelete, antiBanMass, antiKickMass, antiNuke]
    AntiraidConfigInput:
      type: object
      properties:
        enabled:
          type: boolean
        antiAlt:
          type: boolean
        antiAltMinAge:
          type: integer
        antiBot:
          type: boolean
        antiBotWhitelist:
          type: array
          items:
            type: string
        antiSpam:
          type: boolean
        antiSpamLimit:
          type: integer
        antiLinks:
          type: boolean
        allowedDomains:
          type: array
          items:
            type: string
        blockedDomains:
          type: array
          items:
            type: string
        antiMassMention:
          type: boolean
        massMentionLimit:
          type: integer
        antiWebhook:
          type: boolean
        antiChannelCreate:
          type: boolean
        antiChannelDelete:
          type: boolean
        antiChannelUpdate:
          type: boolean
        antiRoleCreate:
          type: boolean
        antiRoleDelete:
          type: boolean
        antiRoleUpdate:
          type: boolean
        antiEmojiCreate:
          type: boolean
        antiEmojiDelete:
          type: boolean
        antiBanMass:
          type: boolean
        antiKickMass:
          type: boolean
        antiNuke:
          type: boolean
    AntiraidStats:
      type: object
      properties:
        guildId:
          type: string
        detectedToday:
          type: integer
        blockedAlt:
          type: integer
        blockedVpn:
          type: integer
        blockedBot:
          type: integer
        blockedSpam:
          type: integer
        totalDetections:
          type: integer
      required: [guildId, detectedToday, blockedAlt, blockedVpn, blockedBot, blockedSpam, totalDetections]
    LogsConfig:
      type: object
      properties:
        id:
          type: integer
        guildId:
          type: string
        enabled:
          type: boolean
        channelId:
          type: ["string", "null"]
        logMembers:
          type: boolean
        logMessages:
          type: boolean
        logRoles:
          type: boolean
        logChannels:
          type: boolean
        logModeration:
          type: boolean
        logSecurity:
          type: boolean
      required: [id, guildId, enabled, logMembers, logMessages, logRoles, logChannels, logModeration, logSecurity]
    LogsConfigInput:
      type: object
      properties:
        enabled:
          type: boolean
        channelId:
          type: ["string", "null"]
        logMembers:
          type: boolean
        logMessages:
          type: boolean
        logRoles:
          type: boolean
        logChannels:
          type: boolean
        logModeration:
          type: boolean
        logSecurity:
          type: boolean
    LogEntry:
      type: object
      properties:
        id:
          type: integer
        guildId:
          type: string
        userId:
          type: ["string", "null"]
        username:
          type: ["string", "null"]
        action:
          type: string
        category:
          type: string
        details:
          type: ["string", "null"]
        moderatorId:
          type: ["string", "null"]
        moderatorName:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, guildId, action, category, createdAt]
    BlacklistEntry:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        username:
          type: string
        reason:
          type: string
        addedBy:
          type: string
        createdAt:
          type: string
      required: [id, userId, username, reason, addedBy, createdAt]
    BlacklistInput:
      type: object
      properties:
        userId:
          type: string
        username:
          type: string
        reason:
          type: string
      required: [userId, username, reason]
    Backup:
      type: object
      properties:
        id:
          type: integer
        guildId:
          type: string
        label:
          type: string
        size:
          type: integer
        version:
          type: integer
        createdAt:
          type: string
      required: [id, guildId, label, size, version, createdAt]
...