# lib/api-zod/src/generated/types/blacklistInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Neuralix Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface BlacklistInput {
  userId: string;
  username: string;
  reason: string;
}